# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Diagnostic Center Billing ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /abnormal-findings/{id}:
    get:
      operationId: getAbnormalFinding
      tags: [abnormal-findings]
      summary: Get an abnormal finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Abnormal finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbnormalFinding"
    patch:
      operationId: updateAbnormalFinding
      tags: [abnormal-findings]
      summary: Update an abnormal finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAbnormalFindingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbnormalFinding"
    delete:
      operationId: deleteAbnormalFinding
      tags: [abnormal-findings]
      summary: Delete an abnormal finding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /abnormal-findings/{id}/use:
    post:
      operationId: incrementAbnormalFindingUse
      tags: [abnormal-findings]
      summary: Increment usage counter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated abnormal finding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbnormalFinding"
  /abnormal-findings:
    get:
      operationId: listAbnormalFindings
      tags: [abnormal-findings]
      summary: List abnormal findings (filtered)
      parameters:
        - name: testId
          in: query
          required: false
          schema:
            type: integer
        - name: modality
          in: query
          required: false
          schema:
            type: string
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            maximum: 500
      responses:
        "200":
          description: Findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbnormalFinding"
    post:
      operationId: createAbnormalFinding
      tags: [abnormal-findings]
      summary: Create an abnormal finding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAbnormalFindingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbnormalFinding"

  # Inventory
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentBase"
...
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment fields or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentBase"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  # Expenses
...
components:
  schemas:
...
    AppointmentPatient:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required: [id, patientId, firstName, lastName, phone]
    AppointmentDoctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    AppointmentBase:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: string
        patientId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        appointmentDate:
          type: string
        timeSlot:
          type: string
        status:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
        ledgerId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, appointmentId, patientId, appointmentDate, timeSlot, status, type]
    Appointment:
      allOf:
        - $ref: "#/components/schemas/AppointmentBase"
        - type: object
          properties:
            patient:
              oneOf:
                - $ref: "#/components/schemas/AppointmentPatient"
                - type: "null"
            doctor:
              oneOf:
                - $ref: "#/components/schemas/AppointmentDoctor"
                - type: "null"
...
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        appointmentDate:
          type: string
        timeSlot:
          type: string
        status:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
      required: [patientId, appointmentDate, timeSlot]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
        timeSlot:
          type: string
        appointmentDate:
          type: string
        doctorId:
          type: ["integer", "null"]
        packageId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        type:
          type: string

    # Expense schemas
...
    AbnormalFinding:
      type: object
      properties:
        id:
          type: integer
        testId:
          type: integer
          nullable: true
        modality:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        keyword:
          type: string
        aliases:
          type: string
          nullable: true
        description:
          type: string
        severity:
          type: string
        isActive:
          type: boolean
        usageCount:
          type: integer
      required: [id, keyword, description]
    CreateAbnormalFindingBody:
      type: object
      properties:
        testId:
          type: integer
          nullable: true
        modality:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        keyword:
          type: string
        aliases:
          type: string
          nullable: true
        description:
          type: string
        severity:
          type: string
        isActive:
          type: boolean
      required: [keyword, description]
    UpdateAbnormalFindingBody:
      type: object
      properties:
        testId:
          type: integer
          nullable: true
        modality:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        keyword:
          type: string
        aliases:
          type: string
          nullable: true
        description:
          type: string
        severity:
          type: string
        isActive:
          type: boolean

    # Inventory
...